Move logging code to separate utility modules
