+++ config/gunicorn.conf.py
@@ -1,5 +1,5 @@
 workers = 2
 bind = "0.0.0.0:8060"
-logger_class = "snekbox.GunicornLogger"
+logger_class = "snekbox.utils.gunicorn.GunicornLogger"
 access_logformat = "%(m)s %(U)s%(q)s %(s)s %(b)s %(L)ss"
 access_logfile = "-"
